Added Jakarta tests to CustomExtensionParsingTests Signed-off-by:Phillip Kruger <phillip.kruger@gmail.com>
